+++ frontends/web/src/components/balance/balance.test.tsx
@@ -45,11 +45,13 @@ describe('components/balance/balance', () => {
           ILS: '512',
           JPY: '512',
           KRW: '512',
+          NGN: '512',
           NOK: '512',
           PLN: '512',
           RUB: '512',
           SEK: '512',
           SGD: '512',
+          ZAR: '512',
           USD: '512',
         }
       }