+++ frontends/web/src/utils/env.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-export const debug = process.env.NODE_ENV === 'development';
+export const debug = import.meta.env.DEV;
 
 /**
  * Returns whether the code is running in QtWebEngine.