+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ## 4.39.0
 - Bundle BitBox02 firmware version v9.15.0
 - Display the wallet root fingerprint in the account info and device settings
+- Add support for Nigerian Naira (NGN)
+- Add support for South African Rand (ZAR)
 
 ## 4.38.0
 - Bundle BitBox02 firmware version v9.14.1