+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Unreleased
+- Drop support for SAI token
+
+## 4.39.0
+- Bundle BitBox02 firmware version v9.15.0
 - Display the wallet root fingerprint in the account info and device settings
 
 ## 4.38.0